check mode for delete operations

+++ plugins/modules/ntnx_foundation_image_upload.py
@@ -96,8 +96,15 @@ def delete_image(module, result):
     image = Image(module, delete_image=True)
     fname = module.params["filename"]
     itype = module.params["installer_type"]
+
+    if module.check_mode:
+        result["file_name"] = fname
+        result["response"] = "Image with name:{0} will be deleted.".format(fname)
+        return
+
     resp = image.delete_image(fname, itype)
 
+    result["file_name"] = fname
     result["changed"] = True
     result["response"] = resp
