+++ plugins/modules/ntnx_ndb_stretched_vlans.py
@@ -186,8 +186,14 @@ def delete_stretched_vlan(module, result):
             msg="stretched_vlan_uuid is required field for delete", **result
         )
 
+    if module.check_mode:
+        result["uuid"] = uuid
+        result["response"] = "Stretched vlan with uuid:{0} will be deleted.".format(uuid)
+        return
+
     resp = stretched_vlan.delete_stretched_vlan(uuid)
 
+    result["uuid"] = uuid
     result["response"] = resp
     result["changed"] = True
