+++ plugins/modules/ntnx_ndb_time_machine_clusters.py
@@ -305,8 +305,17 @@ def delete_data_access_instance(module, result):
         result["error"] = err
         err_msg = "'cluster' is required field for removing cluster from time machine"
         module.fail_json(msg=err_msg, **result)
+
+    if module.check_mode:
+        result["cluster_uuid"] = cluster_uuid
+        result["time_machine_uuid"] = tm_uuid
+        result["response"] = "Cluster with uuid:{0} will be deleted from Time Machine.".format(cluster_uuid)
+        return
+
     resp = tm.delete_data_access_instance(tm_uuid, cluster_uuid)
 
+    result["cluster_uuid"] = cluster_uuid
+    result["time_machine_uuid"] = tm_uuid
     result["response"] = resp
     result["changed"] = True
